+++ src/main/java/com/ojins/chatbot/analyzer/ChineseSynonymAnalyzer.java
@@ -14,7 +14,18 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
- * Created by han on 11/14/16.
+
+ ___   ___  ________  ___   __      __     __   ________ ________  ______
+ /__/\ /__/\/_______/\/__/\ /__/\   /__/\ /__/\ /_______//_______/\/_____/\
+ \::\ \\  \ \::: _  \ \::\_\\  \ \  \ \::\\:.\ \\__.::._\\::: _  \ \:::_ \ \
+ \::\/_\ .\ \::(_)  \ \:. `-\  \ \  \_\::_\:_\/   \::\ \ \::(_)  \ \:\ \ \ \
+ \:: ___::\ \:: __  \ \:. _    \ \   _\/__\_\_/\ _\::\ \_\:: __  \ \:\ \ \ \
+ \: \ \\::\ \:.\ \  \ \. \`-\  \ \  \ \ \ \::\ /__\::\__/\:.\ \  \ \:\_\ \ \
+ \__\/ \::\/\__\/\__\/\__\/ \__\/   \_\/  \__\\________\/\__\/\__\/\_____\/
+
+
+
+ * Created on 11/14/16.
  */
 public class ChineseSynonymAnalyzer extends Analyzer {
     private final CharArraySet stopWords;
