+++ src/test/java/com/ojins/chatbot/testStateIO.java
@@ -1,0 +1,45 @@
+package com.ojins.chatbot;
+
+import com.google.common.collect.Sets;
+import com.ojins.chatbot.dialog.QAPairBuilder;
+import com.ojins.chatbot.dialog.StateIO;
+import lombok.extern.slf4j.Slf4j;
+import lombok.val;
+import org.junit.Assert;
+import org.junit.Test;
+
+import java.io.IOException;
+
+/**
+ ___   ___  ________  ___   __      __     __   ________ ________  ______
+ /__/\ /__/\/_______/\/__/\ /__/\   /__/\ /__/\ /_______//_______/\/_____/\
+ \::\ \\  \ \::: _  \ \::\_\\  \ \  \ \::\\:.\ \\__.::._\\::: _  \ \:::_ \ \
+ \::\/_\ .\ \::(_)  \ \:. `-\  \ \  \_\::_\:_\/   \::\ \ \::(_)  \ \:\ \ \ \
+ \:: ___::\ \:: __  \ \:. _    \ \   _\/__\_\_/\ _\::\ \_\:: __  \ \:\ \ \ \
+ \: \ \\::\ \:.\ \  \ \. \`-\  \ \  \ \ \ \::\ /__\::\__/\:.\ \  \ \:\_\ \ \
+ \__\/ \::\/\__\/\__\/\__\/ \__\/   \_\/  \__\\________\/\__\/\__\/\_____\/
+
+ * Created on 11/14/16.
+ */
+
+@Slf4j
+public class testStateIO {
+
+    @Test
+    public void testRead() throws IOException {
+        val fileName = getClass().getClassLoader().getResource("statedb-test.json").getPath();
+        Assert.assertEquals(1, StateIO.loadStatesFromJson(fileName).size());
+    }
+
+    @Test
+    public void testWrite() throws IOException {
+
+        val qaStates = Sets.newHashSet(
+                new QAPairBuilder().setQuestion("测试问题1").setAnswer("回答").build(),
+                new QAPairBuilder().setQuestion("测试问题2").setAnswer("回答").build());
+
+        val fileName = getClass().getClassLoader().getResource("statedb-test.json").getPath();
+        StateIO.writeStatesToJson(qaStates, fileName);
+        Assert.assertEquals(2, StateIO.loadStatesFromJson(fileName).size());
+    }
+}
