+++ src/test/resources/statedb-test.json
@@ -1,0 +1,6 @@
+[
+  {
+    "question": "今天天气真不错",
+    "answer": "是啊, 太阳特足!"
+  }
+]
