--- src/test/statedb-test.json
@@ -1,12 +1,0 @@
-[
-  {
-    "questions": [
-      "q1",
-      "q2"
-    ],
-    "answers": [
-      "a1",
-      "a2"
-    ]
-  }
-]
